api: disable vanilla cinder-wsgi apache2 conf earlier
If apache is (re)starting while the chef-client installs the cinder-api
package, it may pick up the package-supplied cinder-wsgi apache2 conf
file before chef-client gets around to disabling it which may result in
apache2 failing due to a non-working configuration.

This changeset minimizes the time window for the race condition by
moving the code for disabling the vanilla cinder-wsgi apache2 conf
right after the package installation.

Change-Id: I5e531fac2153166b8f2d02fa3aeb9df3313d6b3f